CuteRobot/pyproject.toml
[tool.poetry]
name = "CuteRobot"
version = "0.1.0"
description = "CuteRobot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_htmlrender", "nonebot_plugin_manager", "nonebot_plugin_logo", "nonebot_plugin_apscheduler"]
plugin_dirs = ["cuterobot/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
